feat: add ability to specify additional_jobs in configuration
Previously, usage of the "checks" array would cause usage of that value for the matrix, and no further discovery would occur.
This patch adds a new key, "additional_checks", which allows package authors to provide checks to run _in addition to any discovered_.
This allows providing checks for tools the matrix discovery tools do not know about, or providing one-off checks (such as benchmarks or mutation tests).

The syntax for the "additional_checks" key is as follows:

```json
{
    "additional_checks": [
        {
            "name": "(string; REQUIRED) name of the check to run",
            "job": {
                "command": "(string; REQUIRED) command to run",
                "php": "(string; OPTIONAL) PHP version to run on (defaults to stable-php); * indicates all versions",
                "dependencies": "(string; OPTIONAL) dependency set to run on (defaults to locked); * indicates all sets",
                "extensions": [
                    "(array of strings; OPTIONAL) specific extensions to install for this check only"
                ],
                "ini": [
                    "(array of strings; OPTIONAL) specific php.ini settings to use for this check only"
                ]
            }
        }
    ]
}
```

A job per PHP version per dependency set will be created, and the "name" will be appended with "on PHP {VERSION} with {DEPS} dependencies" during an actual run.

The tool discovers checks first, then appends any `additional_checks` are concatenated, and then any `exclude` rules are applied.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>