feat: normalize object jobs in "checks" configuration to JSON strings
Adds functionality to validate checks found in the "checks" array, including the ability to validate and normalize the "job" element.
If the "job" element is a string, it attempts to parse it, raising a warning if the job is unparseable, and excluding it from the list of checks returned.
If the "job" is parseable and/or an object, it then validates that a valid PHP version is present, a valid "dependencies" set is specified, and that a command is present and a string.
If valid, the job is serialized to JSON and the value reassigned in the check.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>