feat: allow toggling whether or not to ignore platform reqs on PHP 8
Since the beginning of development on this action, we've used the `--ignore-platform-reqs` flag when installing dependencies on PHP 8.
However, with more packages in the Composer ecosystem supporting PHP 8, we can often safely remove that option and still successfully install packages.
To enable that behavior in a BC way, this patch adds the ability to provide an `ignore_platform_reqs_8` flag via configuration.
By default, the value is `true`, but the configuration allows you to disable it.
Actions working in conjunction with this one can thus opt-in to the configuration option to determine whether or not to toggle that Composer option.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>