fix: ensure pull request branch name is unique
When a pull request is made by a user and they name their branch the same name as the target branch against which the pull request is being made, we run into a conflict, which causes the checkout to fail (see https://github.com/mezzio/mezzio-helpers/pull/7/checks?check_run_id=2393994090#step:3:9).

To resolve this scenario, if we detect a pull_request, we can give the local branch name a prefix, `pull/`.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>