Run `infection` via `phpdbg` by default, since `phpdbg` is pre-installed with PHP
This removes the need for installing `pcov` or `xdebug` in environments that need to
run mutation tests: `phpdbg` being pre-installed means that any bugs with it will
also be reported to the PHP project, instead of dragging in bugs from third-party
extensions.

This also fixes builds for projects that did not declare `ext-pcov` or `ext-xdebug`
as dependencies in `require-dev`: no need to have those, when running this task.