refactoring: remove github event usage and instead target `GITHUB_BASE_REF` environment variable
This also introduces a new tool type `ToolRunningContainerDefaultPhpVersion`. This is used in case tools are pre-installed within the CI container as it is the case for `roave-backward-compatibility-check`.

By having an option available for `backwardCompatibilityCheck`, projects can opt-in for these kind of checks. This flag will be enabled by-default with v2 and thus changes to opt-out.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>